# docs/api/openapi-v1-phase2.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Alter Earth API",
        "description": "A community platform for ecological conservation, sustainability advocacy, and using AI for environmental good",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/posts/{post_id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get Post",
                "description": "Get a single post by ID with full details.",
                "operationId": "get_post_api_v1_posts__post_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/comments/{comment_id}": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Get Comment",
                "description": "Get a single comment by ID.",
                "operationId": "get_comment_api_v1_comments__comment_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "comments"
                ],
                "summary": "Update Comment",
                "description": "Update a comment's content.\n\nOnly the comment author can update it.",
                "operationId": "update_comment_api_v1_comments__comment_id__put",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CommentUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "comments"
                ],
                "summary": "Delete Comment",
                "description": "Soft delete a comment.\n\nOnly the comment author can delete it.\nPreserves thread structure by marking as deleted instead of removing.",
                "operationId": "delete_comment_api_v1_comments__comment_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/votes/posts/{post_id}": {
            "post": {
                "tags": [
                    "votes"
                ],
                "summary": "Vote On Post",
                "description": "Vote on a post.\n\n- vote_value: 1 for upvote, -1 for downvote\n- If user already voted, updates the vote\n- Cannot vote on own posts\n- Vote weight is based on user's karma",
                "operationId": "vote_on_post_api_v1_votes_posts__post_id__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VoteCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "votes"
                ],
                "summary": "Remove Vote From Post",
                "description": "Remove vote from a post.\n\nReturns 204 even if vote didn't exist.",
                "operationId": "remove_vote_from_post_api_v1_votes_posts__post_id__delete",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/votes/posts/{post_id}/status": {
            "get": {
                "tags": [
                    "votes"
                ],
                "summary": "Get Post Vote Status",
                "description": "Check if current user has voted on a post and what their vote is.",
                "operationId": "get_post_vote_status_api_v1_votes_posts__post_id__status_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Post Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteStatusResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/votes/comments/{comment_id}": {
            "post": {
                "tags": [
                    "votes"
                ],
                "summary": "Vote On Comment",
                "description": "Vote on a comment.\n\n- vote_value: 1 for upvote, -1 for downvote\n- If user already voted, updates the vote\n- Cannot vote on own comments\n- Vote weight is based on user's karma",
                "operationId": "vote_on_comment_api_v1_votes_comments__comment_id__post",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VoteCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/votes/comments/{comment_id}/status": {
            "get": {
                "tags": [
                    "votes"
                ],
                "summary": "Get Comment Vote Status",
                "description": "Check if current user has voted on a comment and what their vote is.",
                "operationId": "get_comment_vote_status_api_v1_votes_comments__comment_id__status_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Comment Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteStatusResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CommentResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "content": {
                        "type": "string",
                        "title": "Content"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "User Id"
                    },
                    "post_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Post Id"
                    },
                    "parent_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Parent Id"
                    },
                    "upvotes": {
                        "type": "integer",
                        "title": "Upvotes"
                    },
                    "downvotes": {
                        "type": "integer",
                        "title": "Downvotes"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score"
                    },
                    "is_deleted": {
                        "type": "boolean",
                        "title": "Is Deleted"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "author": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/CommentAuthor"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "user_vote": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Vote"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "content",
                    "user_id",
                    "post_id",
                    "parent_id",
                    "upvotes",
                    "downvotes",
                    "score",
                    "is_deleted",
                    "created_at",
                    "updated_at"
                ],
                "title": "CommentResponse",
                "description": "Schema for comment in responses (includes author info)."
            },
...
            "CommentUpdate": {
                "properties": {
                    "content": {
                        "type": "string",
                        "maxLength": 10000,
                        "minLength": 1,
                        "title": "Content"
                    }
                },
                "type": "object",
                "required": [
                    "content"
                ],
                "title": "CommentUpdate",
                "description": "Schema for updating a comment.",
                "example": {
                    "content": "Updated comment content..."
                }
            },
...
            "EntityTypeEnum": {
                "type": "string",
                "enum": [
                    "automated_article",
                    "user_submission"
                ],
                "title": "EntityTypeEnum",
                "description": "Post entity type for API responses."
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PostResponse": {
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Id"
                    },
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "entity_type": {
                        "$ref": "#/components/schemas/EntityTypeEnum"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "User Id"
                    },
                    "community_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Community Id"
                    },
                    "upvotes": {
                        "type": "integer",
                        "title": "Upvotes"
                    },
                    "downvotes": {
                        "type": "integer",
                        "title": "Downvotes"
                    },
                    "score": {
                        "type": "integer",
                        "title": "Score"
                    },
                    "hot_score": {
                        "type": "number",
                        "title": "Hot Score"
                    },
                    "view_count": {
                        "type": "integer",
                        "title": "View Count"
                    },
                    "comment_count": {
                        "type": "integer",
                        "title": "Comment Count"
                    },
                    "is_active": {
                        "type": "boolean",
                        "title": "Is Active"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "automated_article": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/AutomatedArticleResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "user_submission": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/UserSubmissionResponse"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "user_nickname": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Nickname"
                    },
                    "user_type": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Type"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "title",
                    "entity_type",
                    "user_id",
                    "community_id",
                    "upvotes",
                    "downvotes",
                    "score",
                    "hot_score",
                    "view_count",
                    "comment_count",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "PostResponse",
                "description": "Schema for post in API responses."
            },
...
            "VoteCreate": {
                "properties": {
                    "vote_value": {
                        "type": "integer",
                        "enum": [
                            1,
                            -1
                        ],
                        "title": "Vote Value",
                        "description": "1 for upvote, -1 for downvote"
                    }
                },
                "type": "object",
                "required": [
                    "vote_value"
                ],
                "title": "VoteCreate",
                "description": "Schema for creating or updating a vote.",
                "examples": [
                    {
                        "vote_value": 1
                    },
                    {
                        "vote_value": -1
                    }
                ]
            },
            "VoteResponse": {
                "properties": {
                    "user_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "User Id"
                    },
                    "post_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Post Id"
                    },
                    "comment_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Comment Id"
                    },
                    "vote_value": {
                        "type": "integer",
                        "title": "Vote Value"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "vote_value",
                    "created_at",
                    "updated_at"
                ],
                "title": "VoteResponse",
                "description": "Schema for vote in responses."
            },
            "VoteStatusResponse": {
                "properties": {
                    "has_voted": {
                        "type": "boolean",
                        "title": "Has Voted"
                    },
                    "vote_value": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Vote Value"
                    }
                },
                "type": "object",
                "required": [
                    "has_voted"
                ],
                "title": "VoteStatusResponse",
                "description": "Schema for checking user's vote status on an item.",
                "examples": [
                    {
                        "has_voted": true,
                        "vote_value": 1
                    },
                    {
                        "has_voted": false
                    }
                ]
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}